fix: address Greptile review feedback
