+++ {{ cookiecutter.project_slug }}/frontend/src/controllers/copy_controller.js
@@ -1,11 +1,13 @@
 import { Controller } from "@hotwired/stimulus";
 
+import { copyText } from "../utils/clipboard";
+
 export default class extends Controller {
   static targets = ["source", "label"];
 
   async copy() {
     const text = this.sourceTarget.value || this.sourceTarget.textContent || "";
-    const copied = await this.copyText(text);
+    const copied = await copyText(text);
     const original = this.hasLabelTarget ? this.labelTarget.textContent : null;
 
     if (this.hasLabelTarget) {
@@ -14,33 +16,6 @@ export default class extends Controller {
       this.labelTarget.dataset.resetTimer = window.setTimeout(() => {
         this.labelTarget.textContent = original;
       }, 1600);
-    }
-  }
-
-  async copyText(text) {
-    if (navigator.clipboard?.writeText) {
-      try {
-        await navigator.clipboard.writeText(text);
-        return true;
-      } catch (error) {
-        // Fall back for restricted clipboard contexts.
-      }
-    }
-
-    const textarea = document.createElement("textarea");
-    textarea.value = text;
-    textarea.setAttribute("readonly", "");
-    textarea.style.position = "fixed";
-    textarea.style.top = "-9999px";
-    document.body.appendChild(textarea);
-    textarea.select();
-
-    try {
-      return document.execCommand("copy");
-    } catch (error) {
-      return false;
-    } finally {
-      textarea.remove();
     }
   }
 }
